+++ .gitignore
@@ -1,3 +1,3 @@
-
 .idea/
 target/
+scheduler/airflow-home/logs/
