+++ python/insight/config/config.ini
@@ -2,8 +2,8 @@
 prefix = /api/o11y/insight
 
 [DB]
-URL = localhost
-USERNAME = root
+URL = 192.168.170.229
+USERNAME = mcmp
 PASSWORD = 1234
 DATABASE = mcmp
